Pulled cleanData out into a Helper file and then changed everything to use that in LineChart, BarChart, Sparkline, and Pie
