+++ packages/chart/index.html
@@ -44,10 +44,10 @@
     <!-- DATA PRESENTATION GALLERY: https://www.cdc.gov/wcms/4.0/cdc-wp/data-presentation/bar-chart.html#examples -->
 
     <!-- NONNUMERIC TESTS -->
-    <!-- <div class="react-container" data-config="/examples/planet-pie-example-config-nonnumeric.json"></div> -->
+    <div class="react-container" data-config="/examples/planet-pie-example-config-nonnumeric.json"></div>
     <!-- <div class="react-container" data-config="/examples/example-combo-bar-nonnumeric.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/example-bar-chart-nonnumeric.json"></div> -->
-    <div class="react-container" data-config="/examples/line-chart-nonnumeric.json"></div>
+    <!-- <div class="react-container" data-config="/examples/line-chart-nonnumeric.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/sparkline-chart-nonnumeric.json"></div> -->
 
     <!-- HORIZONTAL BAR CHARTS -->
